Restructuring exporter code to pick exporters based on configuration
