+++ delfin/exporter/kafka/exporter.py
@@ -1,0 +1,27 @@
+# Copyright 2020 The SODA Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from delfin.exporter import base_exporter
+from delfin.exporter.kafka import kafka
+
+
+class AlertExporterKafka(base_exporter.BaseExporter):
+    def dispatch(self, ctxt, data):
+        pass
+
+
+class PerformanceExporterKafka(base_exporter.BaseExporter):
+    def dispatch(self, ctxt, data):
+        kafka_obj = kafka.KafkaExporter()
+        kafka_obj.push_to_kafka(data)
