+++ etc/delfin/delfin.conf
@@ -2,6 +2,9 @@
 api_paste_config = /etc/delfin/api-paste.ini
 delfin_cryptor = delfin.cryptor._Base64
 api_max_limit = 1000
+# Uncomment or add exporters
+# performance_exporters = PerformanceExporterPrometheus, PerformanceExporterKafka
+# alert_exporters = AlertExporterPrometheus
 
 [database]
 connection = sqlite:////var/lib/delfin/delfin.sqlite
@@ -11,13 +14,11 @@ db_backend = sqlalchemy
 config_path = /etc/delfin/scheduler_config.json
 
 [KAFKA_EXPORTER]
-enable = False
 kafka_topic_name = "delfin-kafka"
 kafka_ip = 'localhost'
 kafka_port = '9092'
 
 [PROMETHEUS_EXPORTER]
-enable = False
 metric_server_ip = 0.0.0.0
 metric_server_port = 8195
 metrics_cache_file = /var/lib/delfin/delfin_exporter.txt
